Consolidated shared auxillary composition entity insert functions into one file to reduce code duplication
