+++ frontend/src/pages/EditCompositionPage.jsx
@@ -3,6 +3,7 @@ import React, { useState } from 'react';
 import { useNavigate } from "react-router-dom";
 import { server_url } from '../config';
 import { CompositionForm } from '../components/forms/CompositionForm.mjs';
+import * as service from '../modules/compositionService.mjs';
 
 export const EditCompositionPage = ({ compositionToEdit }) => {
   // State variables for database entities
@@ -55,7 +56,7 @@ export const EditCompositionPage = ({ compositionToEdit }) => {
     if (opusNums.length > 0) {
       const response = await fetch(`${server_url}/api/opus-nums/for-composition-${compositionToEdit.compositionID}`, { method: 'DELETE'});
       if (response.ok) {
-        addOpusNums(compositionToEdit.compositionID);
+        service.addOpusNums(compositionToEdit.compositionID, opusNums);
       } else {
         setEditSuccess(false);
         console.error(`Unable to delete opus num(s) from composition with ID ${compositionToEdit.compositionID}, status code = ${response.status}`)
@@ -66,7 +67,7 @@ export const EditCompositionPage = ({ compositionToEdit }) => {
     if (catalogueNums.length > 0) {
       const response = await fetch(`${server_url}/api/catalogue-nums/for-composition-${compositionToEdit.compositionID}`, { method: 'DELETE'});
       if (response.ok) {
-        addCatalogueNums(compositionToEdit.compositionID);
+        service.addCatalogueNums(compositionToEdit.compositionID, catalogueNums);
       } else {
         setEditSuccess(false);
         console.error(`Unable to delete catalogue num(s) from composition with ID ${compositionToEdit.compositionID}, status code = ${response.status}`)
@@ -77,7 +78,7 @@ export const EditCompositionPage = ({ compositionToEdit }) => {
     if (featuredInstrumentation.length > 0) {
       const response = await fetch(`${server_url}/api/featured-instruments/for-composition-${compositionToEdit.compositionID}`, { method: 'DELETE'});
       if (response.ok) {
-        await addFeaturedInstrumentation(compositionToEdit.compositionID);
+        await service.addFeaturedInstrumentation(compositionToEdit.compositionID, featuredInstrumentation);
       } else {
         setEditSuccess(false);
         console.error(`Unable to delete featured instrument(s) from composition with ID ${compositionToEdit.compositionID}, status code = ${response.status}`)
@@ -85,10 +86,10 @@ export const EditCompositionPage = ({ compositionToEdit }) => {
     }
 
     // INSERT Movement(s)
-    if (featuredInstrumentation.length > 0) {
+    if (movements.length > 0) {
       const response = await fetch(`${server_url}/api/movements/for-composition-${compositionToEdit.compositionID}`, { method: 'DELETE'});
       if (response.ok) {
-        await addMovements(compositionToEdit.compositionID);
+        await service.addMovements(compositionToEdit.compositionID, movements);
       } else {
         setEditSuccess(false);
         console.error(`Unable to delete movement(s) from composition with ID ${compositionToEdit.compositionID}, status code = ${response.status}`)
@@ -99,105 +100,6 @@ export const EditCompositionPage = ({ compositionToEdit }) => {
       redirect(`/composition/${compositionToEdit.compositionID}`);
     }
   };
-
-  // Prepare the opus numbers as query compatible objects and send to database
-  const addOpusNums = async (compositionID) => {
-    const data = opusNums.map(opusNum => ({ 
-      compositionID: compositionID, 
-      opNum: opusNum 
-    }));
-    const response = await fetch(`${server_url}/api/opus-nums/for-composition-${compositionID}`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    })
-    if(response.ok){
-      console.log(`Opus number(s) successfully added for composition with ID ${compositionID}!`);
-    } 
-    else {
-      setEditSuccess(false);
-      console.log(`Unable to add opus number(s). Request returned status code ${response.status}`);
-    }
-  }
-
-  // Prepare the catalogue numbers as query compatible objects and send to database
-  const addCatalogueNums = async (compositionID) => {
-    const data = catalogueNums.map(catNum => ({ 
-      catalogueID: catNum.catalogueID, 
-      compositionID: compositionID, 
-      catNum: catNum.catNum 
-    }));
-    const response = await fetch(`${server_url}/api/catalogue-nums/for-composition-${compositionID}`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    })
-    if(response.ok){
-      console.log(`Catalogue number(s) successfully added for composition with ID ${compositionID}!`);
-    } 
-    else {
-      setEditSuccess(false);
-      console.log(`Unable to add catalogue number(s). Request returned status code ${response.status}`);
-    }
-  }
-
-  // Prepare the featured instrumentation as query compatible objects and send to database
-  const addFeaturedInstrumentation = async (compositionID) => {
-    const data = featuredInstrumentation.map(instrument => ({ 
-      compositionID: compositionID,
-      instrumentID: instrument.id
-    }));
-    const response = await fetch(`${server_url}/api/featured-instruments/for-composition-${compositionID}`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    })
-    if(response.ok){
-      console.log(`Featured instrument(s) successfully added for composition with ID ${compositionID}!`);
-    } 
-    else {
-      setEditSuccess(false);
-      console.log(`Unable to add featured instrument(s). Request returned status code ${response.status}`);
-    }
-  }
-
-  // Prepare the movements as query compatible objects and send to database
-  const addMovements = async (compositionID) => {
-    const data = movements.length > 0 ?        
-      // Multi-movement works
-      movements.map(movement => ({ 
-        compositionID: compositionID, 
-        movementNum: movement.num,
-        title:  movement.title === '' ? null : movement.title
-      }))
-    :
-      // Single-movement works
-      [{
-        compositionID: compositionID,
-        movementNum: 1,
-        title: null
-      }]
-    const response = await fetch(`${server_url}/api/movements/for-composition-${compositionID}`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    })
-    if(response.ok){
-      console.log(`Movement(s) successfully added for composition with ID ${compositionID}!`);
-    } 
-    else {
-      setEditSuccess(false);
-      console.log(`Unable to movement(s). Request returned status code ${response.status}`);
-    }
-  }
 
   return (
     <>
